moved maketorrent to a more sensisble place
